requirements.txt
asgiref==3.8.1
attrs==24.2.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.3.2        
Django==4.2.16
djangorestframework==3.15.2      
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
django-bootstrap-v5==1.0.11
django-cleanup==9.0.0
Markdown==3.7
django-filter==24.3
Pygments==2.19.1